requirements-api.txt
# FastAPI Backend Requirements
# Minimal dependencies for the Chaos Engineering API

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.6.0

# HTTP Client
requests>=2.28.0

# AWS Bedrock
boto3>=1.28.0
botocore>=1.31.0

# Environment Variables
python-dotenv>=1.0.0
